Tests done

+++ Kemo_Allen_P0/src/main/java/com/revature/repository/ShopDAOImpl.java
@@ -409,7 +409,7 @@ public class ShopDAOImpl implements ShopDAO{
 		try (Connection conn = ShopConnection.getConnection()){
 			
 			PreparedStatement ps = conn.prepareStatement(sql);
-			ps.setInt(1, id);
+			ps.setInt(1, (int) id);
 			
 			ResultSet rs = ps.executeQuery();
